Revert "tangible-surfaces 구현"
