+++ index.js
@@ -1,0 +1,3 @@
+import MeaningfulMotion from './sub-containers/meaningful-motion.js';
+
+new MeaningfulMotion();
